# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: بذور Seeds Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /upload/image:
    post:
      operationId: uploadImage
      tags: [upload]
      summary: Upload product image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /upload/bulk:
    post:
      operationId: bulkUploadProducts
      tags: [upload]
      summary: Bulk upload products via Excel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Products imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, customer]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
      required:
        - url
    BulkUploadResponse:
      type: object
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - failed
...